requirements.txt
# RAG Chatbot - Python Dependencies
# Project: RAG Chatbot
# Developer: RSK World
# Website: https://rskworld.in
# Email: help@rskworld.in
# Phone: +91 93305 39277
# Year: 2026

flask==3.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
openai==1.10.0
python-dotenv==1.0.0
pypdf2==3.0.1
tiktoken==0.5.2
numpy==1.26.3

